% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundx.R
\name{roundx}
\alias{roundx}
\title{A rounding function}
\usage{
roundx(n, dp)
}
\arguments{
\item{n}{a number to round}

\item{dp}{decimal place to round to}
}
\value{
rounded number
}
\description{
A rounding function
}
\examples{
roundx(10.111, 1)
}
