% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundx.R
\name{roundx_df}
\alias{roundx_df}
\title{Applying the rounding function to all numeric values of dataframe while excluding characters}
\usage{
roundx_df(df, dp)
}
\arguments{
\item{df}{a dataframe}

\item{dp}{decimal place to round to}
}
\value{
a dataframe
}
\description{
Applying the rounding function to all numeric values of dataframe while excluding characters
}
\examples{
roundx_df(data.frame(c(1.111, 1, 1.11), c(1, 1, 1), c("a", "b", "c")), 1)
}
